requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
httpx>=0.28.0
playwright>=1.49.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0

